refactor(table): extract data overlay model into format::overlay module
The overlay data model (`OverlayCoverage`, `DataOverlayFile`, their serde/
protobuf/`DeepSizeOf` impls, the Roaring (de)serialization helpers, and
`sort_overlays_newest_last`) lived interleaved with fragment code in
`fragment.rs`. Move it to its own `format/overlay.rs` submodule so the small
external interface (two types, `dense`/`sparse`, `coverage_for_field`) is
visible and the serde/proto/rank machinery is hidden as module-private
implementation. A module doc consolidates the coverage, rank, parse-once, and
newest-last ordering invariants that were previously scattered across item docs.

Pure code move: `format::DataOverlayFile`/`OverlayCoverage` paths are preserved
via re-export, so callers are unchanged. Pure-overlay unit tests move with the
module; the fragment-carrier round-trip/sort tests stay in `fragment.rs`.
